Small refactor, moved execution of the tester to index.js file
