+++ src/cli.js
@@ -1,13 +1,8 @@
 'use strict';
-const figlet = require('figlet');
-const chalk = require('chalk');
 const path = require('path');
 const meow = require('meow');
 const updateNotifier = require('update-notifier');
-const SpeedTest = require('./tester');
-const reporter = require('./reporter');
-const colors = require('./config/colors');
-const log = console.log;
+const performanceTester = require('./');
 
 
 const cli = meow(`
@@ -35,45 +30,7 @@ updateNotifier({pkg: cli.pkg}).notify();
 
 //  log(`This is how your cli looks --> cli.input[0]: ${cli.input[0]} and cli.flags: ${JSON.stringify(cli.flags)}`);
 
-const defaultTests = cli.input[0] ? require(path.resolve(cli.input[0])) : require('./default-tests/default-tests');
+const listOfTests = cli.input[0] ? require(path.resolve(cli.input[0])) : require('./default-tests/default-tests');
 
 
-/*
- *  We create the speedTest using its constructor for the implementation that we
- *  want to test
- */
-let listOfTest = [];
-
-for ( let prop in defaultTests) {
-  if (defaultTests.hasOwnProperty(prop) ) {
-    listOfTest.push(new SpeedTest(defaultTests[prop]));
-  }
-}
-
-
-figlet('TesterJS', function(err, data) {
-  if (err) {
-    log(chalk.bold.red('Something went wrong...'));
-    console.dir(err);
-    return;
-  }
-
-  executeTester();
-});
-
-/**
- * Main function execute all tests
- */
-function executeTester() {
-  let i = 0;
-
-  log(colors.todo(
-    figlet.textSync('TesterJS')
-  ));
-
-  log(colors.log('TEST RESULTS'));
-
-  for (; i < listOfTest.length; i += 1) {
-    log('\n' + reporter.generateOutputStr(listOfTest[i].executor()));
-  }
-}
+performanceTester.output(listOfTests);
